packages/backend/requirements/base.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Redis Cache
redis==5.0.1
hiredis==2.3.2

# Database (Optional)
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0

# AI Services
openai==1.9.0
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
pendulum==3.0.0
tenacity==8.2.3

# Logging
python-json-logger==2.0.7
structlog==24.1.0

# Monitoring (Optional)
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.2

# Rate Limiting
slowapi==0.1.9

# Security Headers
secure==0.3.0